refactor motor to be component
