+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Components/MotorGroup.kt
@@ -1,0 +1,14 @@
+package org.firstinspires.ftc.teamcode.Components
+
+import org.firstinspires.ftc.teamcode.Components.Motor
+
+public class MotorGroup(){
+    var motors:List<Motor> = listOf()
+
+
+    fun addMotor(motor: Motor){
+        motors.ad
+    }
+
+
+}
